Did some refactoring today as well as some attempts at bug hunting. Updated everything to use python 3.3 and boost 1.55 so if you are using a different version of boost or python you will need to change these settings again (sorry). Refactored the layout of the repository a little bit and changed the exe properties to look for buddybots folder instead of CTFBots now that the project has a name.
Bug Squashing report:
I hate fopen! I'm not sure what is happening but when it goes to open the python script file fopen returns a invalid FILE*, which
would be fine if the function actually produced a error result. According to fopen and fopen_s the file opens just fine.