Updated BuddyBots to compile and run with Visual Studio 2015:
Modified string literal locations to have a space between a string and the following literal value afterwards. This created an error when attempting to build in 2015.
Updated FileSystem to use _fileno() to acquire a file handle rather than attempting to access the internals of the FILE structure. This was a breaking change in visual studio 2015.
Updated documentation on how to build to reference the new Visual studio version and toolset linking for 2015.

To Do in terms of priority:
FIND TESTERS OTHER THAN MYSELF (would like this not to be students at first)
Produce bots that attempt to accurately shoot one another based on information gathered through scripts. (This may be easier than it sounds, there is already code for the monsters to predict where they need to aim, but I have yet to make that available to the bots)
Make sure that bots can travel on elevators in maps correctly. (This is going to be more difficult, I am currently using old mod code that attempts to solve this problem, but it is buggy and will most likely need to be looked at again)
Create better documentation for actually building bots and the API available to users.